When converting the inputs to tensor the dtype should automatically be inferred from the input rather the using the layer dtype as those can be different when doing mixed precision training.
PiperOrigin-RevId: 217385423

+++ tensorflow/python/keras/layers/core.py
@@ -956,7 +956,7 @@ class Dense(Layer):
     self.built = True
 
   def call(self, inputs):
-    inputs = ops.convert_to_tensor(inputs, dtype=self.dtype)
+    inputs = ops.convert_to_tensor(inputs)
     rank = common_shapes.rank(inputs)
     if rank > 2:
       # Broadcasting is required for the inputs.
